[BigQuery] Remove BigQueryTableChecker pattern (Recidiviz/recidiviz-data#38524)
## Description of the change

In the beginning times we used to update views as part of the deploy
which always happened *before* the `state` dataset table schemas were
updated as part of the CloudSQL -> BQ refresh (post-deploy). This meant
that the views that are derived from the schemas of those tables always
would crash at deploy time when a new column was added, because we
wouldn't have update the source table schemas yet.

To solve this problem, we added a predicate to a number of views that,
at build time, would QUERY ACTUAL LIVE BQ to see if the appropriate
column existed and we could create the view. If it didn't exist, we'd
wait until the next view update cycle to create it.

A lot has changed since then. We now update source table schemas and
views as part of the calc DAG and this always happens in a predictable
order.

This PR removes the `has_column_predicate` checks from all views and
deletes the `BigQueryTableChecker`, which is now unused. This will,
among other things, speed up `load_views_to_sandbox` `auto` runs, since
we call `should_deploy()` from there as well.

In a follow-up PR, I will actually remove all the
`should_deploy_predicate` / `should_deploy()` infra since it is now all
a no-op (always returns True). This is *different* than the
`should_deploy_in_project` functionality which lives on
`BigQueryViewBuilder` and remains intact.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Important for load_views_to_sandbox perf work.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b75f02cde14dfd6d4de90ed498e15a56973fdbe6