[US_TN] adding cte to make sure LastUpdateDate is updated with changes (Recidiviz/recidiviz-data#38521)
## Description of the change

TN sent us data where Sentence information changed but their
LastUpdateDate did not which led us to have two expiration dates for the
same LastUpdateDate in task deadlines and ingest failed. This way we
make sure to get the latest data for the same LastUpdateDate, for this
instance and for if it occurs in the future.

Successful prod ingest pipeline
[here](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-02-20_10_35_08-14612522572466922863;graphView=0?project=recidiviz-123&authuser=0&inv=1&invt=AbqGfA).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6249b2a645c28b8682bc4a47a46958d0a0b9bd4e