[US_NE] Ingesting sentence status snapshot  (Recidiviz/recidiviz-data#38145)
## Description of the change

Adding in sentence status snapshot to NE. There is no sentence status
data directly in Nebraska, so we have to determine the status by looking
at the latest movement for each person from LOCATION_HIST table. If the
last period's movement has an end date, we assume their sentence is
completed. We only know *why* someone's sentence was completed if they
went directly from facility -> liberty because there is only a reason
populated for institution releases, not for parole to liberty.

Succesful ingest pipeline
[here](https://console.cloud.google.com/dataflow/jobs/us-west1/2025-02-07_11_35_26-15716894163363131543?authuser=0&inv=1&invt=Abo9Xg&organizationId=448885369991&project=recidiviz-staging).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#37977

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: dff24f5e8213233e4a57c8e502986603329ea4a9