[BigQuery] Remove should_deploy_predicate pattern from BigQueryView (Recidiviz/recidiviz-data#38525)
## Description of the change

With the changes in Recidiviz/recidiviz-data#38524, the `should_deploy_predicate` on
`BigQueryViewBuilders` is always set to `None`, so `should_deploy()` is
always `True`. This PR removes the `should_deploy_predicate` /
`should_deploy()` pattern entirely since it's a no-op and easily
confused with the `should_deploy_in_project()` functionality on the
`BigQueryViewBuilder`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Important for load_views_to_sandbox perf work.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 969bba8fe22bf176f8c6f3c175ff35134085f691