[Build] Lazily import nltk library to improve import performance (Recidiviz/recidiviz-data#38537)
## Description of the change

I profiled the imports for `load_views_to_sandbox`, which always seem to
take a long time. The imports take 3.7 seconds and .7 of that time is
spent importing the `nltk` package.

<img width="1642" alt="image"
src="https://github.com/user-attachments/assets/f3c773c2-237e-414c-91dc-f97178d05caa"
/>

This PR updates to defer those imports so that we only import that
package in the context of the IX supplemental pipelines that use that
package. With this change, imports are now down to just under 3s when I
run them locally.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A - quick wins week optimization

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7638140ae8f61a93107476bfe13fc52b4d34ff6b