[Ingest Mappings] EntityTreeManifest support for non-normalized JSON nodes (Recidiviz/recidiviz-data#38589)
## Description of the change

This is the second of three PRs that will ultimately address Recidiviz/recidiviz-data#38554.
This PR refactors the `NormalizedJSON` into a `SerializableJSON` class
and `NormalizedSerializableJSON` subclass, then updates the
`SerializedJSONDictFieldManifest` to choose which serialization
mechanism to use based on a new `normalize_values` flag.

This PR does not actually make it possible to add non-normalized JSON in
mappings YAML files - that will come in the next PR and this PR has
added some TODOs for that work.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#38554

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 17389857139ee28879b00c87a4993d9abc418262