[JII Texts] Add jii_to_text product view to union all topic views; export only this view (Recidiviz/recidiviz-data#38583)
## Description of the change

This PR creates the `jii_to_text` product view that will be the file
exported to GCS.

As we grow the number of topics and states, each new topic should be
unioned into this view. Then, the export logic will filter for the state
when exporting the BQ view to GCS, so the `jii_to_text.json` file will
have all people for all topics for a given state. This makes the ETL
more straightforward so that as we have multiple topics, we still have
one file that hydrates one table

Loaded into sandbox with prefix `alexa`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-dashboards/issues/7263

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 12d21056e834b4519ed2f0d707ef2ac406a0cf4a