[US_UT] Early Termination criteria analysis (Recidiviz/recidiviz-data#38364)
## Description of the change

I made a bunch of changes to our Early Termination task query here.
Specifically, I created 6 additional criteria based on our analysis:

1.
[at_least_3_months_since_most_recent_positive_drug_test.py](https://github.com/Recidiviz/recidiviz-data/pull/38364/files#diff-3cbb60c389947e8f7cce702bca77f102abb03250537415ee37008b79e597a3f5):
I re-used some infra for this, but I made adjustments to the helper
function so that it is a bit more efficient and it could be used for
more cases (e.g. any drug screen instead of just positive ones, this
will be helpful for an additional criteria I will try to build for this
opp/).
2.
[on_supervision_at_least_6_months.py](https://github.com/Recidiviz/recidiviz-data/pull/38364/files#diff-4175ea592918a8280a46642709e1273e0ff76c5e650e404d613e72bc23466f94):
I re-used a helper function.
3.
[supervision_continuous_employment_for_3_months.py](https://github.com/Recidiviz/recidiviz-data/pull/38364/files#diff-22507d71f77606f411651cc6a43a95e36413a526ea58f43cc83d71ffcff2375b):
I created a helper function for this one, which is ultimately based on
`status_for_at_least_x_time_criteria`
4.
[supervision_housing_is_permanent_or_temporary_for_3_months.py](https://github.com/Recidiviz/recidiviz-data/pull/38364/files#diff-58b859822f7c7f6fd6cb5ac5d071a06a308ae2c2035db9c0cca09cbfb3b8f227):
same as 3.
5.
[no_medhigh_supervision_violation_within_3_months.py](https://github.com/Recidiviz/recidiviz-data/pull/38364/files#diff-aac26a4a7a5827245f620ca14d29f49ac15478a152ecf2e11207212a603afc7e):
I reused a general criteria builder function, but I had to repurpose the
helper function to recover the actual query from it. You'll find this
change in the helper function here:
[general_criteria_builders.py](https://github.com/Recidiviz/recidiviz-data/pull/38364/files#diff-142afa4a21dfc1d774c8b6bce5304804f69ea2e86fd1ec606b1333e23f27da3b)
6.
[no_risk_level_increase_of_15_percent.py](https://github.com/Recidiviz/recidiviz-data/pull/38364/files#diff-44843c28d2c9782814f2eb58fa37a0b40c5f36238742eb1d357138a4c66aa04d):
this is probably the only query that was written in the actual criteria
py file.

I would really appreciate any naming suggestions for our new
criteria/functions!

I uploaded it all to a sandbox and it's looking well:
```
python -m recidiviz.tools.load_views_to_sandbox \
        --project_id recidiviz-staging\
        --sandbox_dataset_prefix hsalas\
        --state_code_filter US_UT auto\
        --load_changed_views_only
```

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8c75ffdc6fd5b6bcf04cbf322a141c4a16dbbd07