[Cleanup] remove UsTnCompliantReportingWorkflowsImpact table (Recidiviz/recidiviz-data#38595)
## Description of the change

Removes the DB table for the impact dashboard. This table was populated
by ETL, so it isn't managed by alembic- I tried writing a migration to
delete it manually, but that caused some tests to fail since our infra
thinks the table doesn't exist and therefore can't be dropped, so I'm
just going to set a reminder for myself to log into Cloud SQL and delete
the table after this PR is deployed.

The other two relevant PRs
(https://github.com/Recidiviz/recidiviz-data/pull/38540,
https://github.com/Recidiviz/recidiviz-data/pull/38545) made it into
today's rc, so this is safe to merge whenever (since it'll be in a
different staging deploy and a different prod deploy from the other
two).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-dashboards/issues/5045

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1b8c3dfc85fad93427e639a85a89aec9abc2c90e