[US_IX][Workflows] Update release district to pull from case notes first, than jurisdiction (Recidiviz/recidiviz-data#38544)
## Description of the change

Idaho originally wanted us to sort by the CRC Release district case
note, but it has historically been poorly hydrated. We therefore decided
to pull from jurisdiction, however this is causing issues since
residents may request to be transferred to a CRC that is not where their
current jurisdiction is for a variety of reasons. More details in the
issue Recidiviz/recidiviz-data#38445 .

Here we default to the release case note and if that is null, use
jurisdiction. Unfortunately, for most residents both fields are still
null. Validation
[here](https://docs.google.com/spreadsheets/d/1r98botveckWd_8-4Tf7lod063iRmn8_WvMuDBaj6a8c/edit?gid=153243105#gid=153243105).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38445
## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d0bc92eeffcf301178ce8be4ce6e41c4ceee2bba