[Justice Counts] Enable Standardization and Bulk Upload in Single-Page Format (Recidiviz/recidiviz-data#38332)
This PR allows us to standardize and upload a sheet without converting
it to multi-sheet format.

Why do this right now?

Bulk Upload has been slow and memory-intensive, and we’ve even run out
of memory in the past. To fix this, the original plan was to speed
things up by chunking the data in two places:

1) Chunking in WorkbookStandardizer, then concatenating everything back
together.
2) Chunking again in WorkbookUploader before uploading to the database.

But that’s not very efficient. Concatenating chunks brings back high
memory usage, which defeats the purpose of chunking in the first place.
We’re also reading the data into memory multiple times, which slows
everything down.

Instead of loading, chunking, concatenating, and re-chunking, we should
just stream the data straight into the database in chunks as we process
it. This way, we can keep memory usage low by processing chunks one at a
time and avoid unnecessary operations like concatenation that slow
things down.

Previously, handling the single-page format involved opening the file,
reading through all the rows, and reformatting it into the multi-page
format. After that, we standardized the Excel workbook (checking for
incorrect column names, values, etc.) and passed the multi-page format
to WorkbookUploader for ingestion.

This approach prevents us from implementing streaming because it
requires loading the entire file into memory for reformatting—precisely
the issue this refactor aims to eliminate. As a result, this PR removes
the reformatting logic for single-page uploads and ensures we can
standardize and process the single-page format directly without that
extra step.

I've thoroughly tested this change locally.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: bc0c55256db81c6107caebfdd76d0e9ebb807606