[Sentencing][US_ND] Update Case counties to use names instead of numbers (Recidiviz/recidiviz-data#38745)
## Description of the change
The raw docstars offense table uses numbers for counties instead of
there names, so we do an extra join on the location metadata table to
convert them to names.

In order for this to work, we need to update the countyToDistrictId in
the lcoation metadata table to be of the format `COUNTY-X`, where X is
the county number.

Tested this out in a sandbox and it works out of the box because the
countyToDistrictIds are already match the county numbering (what luck!)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fc2c916bcd9e23e3b5d9f0d220b08ff699bd1080