[Ingest] Deprecate IngestViewContentsContext (Recidiviz/recidiviz-data#38579)
## Description of the change

This PR:
- moves `launchable_ingest_views` from the ingest mapping collector to
the ingest view collector
- deprecates `launch_env` from ingest mappings and adds
`allowed_project_ids` and `can_run_in_project` to the ingest view
builder
- deprecates `IngestViewContentsContext` and the `EnvManifestNode` from
parsers

In short we are moving the gating of ingest views from this context to
the `project_id`

This is a pretty big change, so there are three commits:
1. Add functionality to the ingest view and collector / update the
launchable ingest view call
2. Delete the now unused stuff and update existing mappings (most are in
OZ)
3. Add a new test to the ingest view builder

This created one behavior change: dev-only ingest views now _do not run_
in the integration tests, which I feel is the correct behavior. They
still require an ingest view test. If the ingest view is turned off
though, I do not think we should be checking how those results merge up
with other views in the integration test. Interestingly, the AZ sentence
length view was still running in staging despite being "turned off".
This view no longer producing results in the integration test makes me
more confident in this change.

Specific feedback I'm interested in:
- Do y'all like the list of project ids as an option? I feel like if we
are gating, we're basically only gating to a single project OR just
going to dev mode. I considered only having a `development` boolean
instead of projects.
- I deleted the test that is using
`INGEST_VIEW_TREE_MERGER_ERROR_EXEMPTIONS` because it seemed to only be
checking that the views in that list are running? I can update it if I
am just misunderstanding the purpose.
- Let me know if I went overboard in ripping things out. Once deciding
the ingest view collector was responsible for finding launchable views
we were able to rip out a lot of things!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#37799

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1c3f68d1f38bb016330a10f095f73ebb59effaba